Crate leads

LEADS Crate

LEADS is a Lazy Exploratory Analysis Data Summarizer.

Writing the same boilerplate exploratory analysis code in a Juptyer notebook or Excel spreadsheet for each new dataset can be tedious. This crate offers functionality for generating automatic exploratory analysis results for a dataset.

Quickstart

TODO

Direct Dependencies

chrono-0.4.38 used for date and time functionality.

clap-4.5.8 used for command line argument handling when run in comand line mode.

colored-2.1.0 used for colored terminal text when run in command line mode.

indexmap-2.3.0 used for ordered hash tables for processing data columns in a consistent order.

indicatif-0.17.8 used for progress bar functionality when run in command line mode.

pdfium-render-0.8.22 used for Rust bindings to Pdfium for creating PDF reports.

polars-0.41.3 used for performing operations on the dataset.

Opt-in features:

polars-lazy-0.41.3 for the Polars lazy API.

polars-parquet-0.41.3 for support for reading parquet files.

moment for kurtosis and skew statistics.

dtype-array for array data types.

random for random sampling of the dataset.

thiserror-1.0.63 for defining library errors.

plotters-0.3.7 for generating visualizations.

Modules

data

prelude

report

spinner

Spinner Module

Enums

LeadsError

High level errors in the LEADS crate.

Type Aliases

LeadsResult

LEADS result type alias, used for handling results throughout the crate.