Module leads::data::base

Base Data Module

This module handles loading data into a Polars LazyFrame from various file formats. It provides functionality to read CSV, TSV, and Parquet files, and performs initial data processing and analysis.

TODO : clean this up

Examples

Structs

DataInfo

Struct to hold the data information, analysis results, and analysis metadata.

Enums

DataError

The error types for the base data module.