Module leads::report::pdf

PDF Module

Handles the base implementation of generating a comprehensive PDF report with the exploratory analysis findings.

Structs

PageManager

Struct that keeps track of the current page position and number. Allows for manual page management, page break handling, and flow content across multiple pages.

Enums

PdfError

The error types for the pdf modules.

Constants

BOLD_FONT

The default bold font.

BOTTOM_MARGIN

Bottom page margin.

FEATURE_HEADER_FONT_SIZE

Sub-header for feature names.

FONT

The default font.

FONT_SIZE

Normal text font size.

ITALIC_FONT

The default italic font.

LINE_HEIGHT_PADDING

Padding between normal lines of text.

PAPER_SIZE

The default paper size.

SECTION_HEADER_FONT_SIZE

Section header font size.