Struct leads::report::pdf::PageManager

pub struct PageManager<'a> { /* private fields */ }

Struct that keeps track of the current page position and number. Allows for manual page management, page break handling, and flow content across multiple pages.

Implementations

impl<'a> PageManager<'a>

pub fn new(pdfium: &'a Pdfium) -> Result<Self, PdfError>

Constructor for the PageManager struct.

Parameters

pdfium: Reference to a Pdfium struct.

Returns

PageManager: The new PageManager.

pub fn generate_report( &mut self, data_info: &DataInfo, ) -> Result<(), LeadsError>

Generates the final report.

Parameters

data_info: The dataset information.

pub fn create_title_page(&mut self, data_title: &str) -> Result<(), LeadsError>

Create the report title page.

Parameters

data_title: Title of the dataset.

Returns

Result<(), PdfError>: Unit type or a propagated PdfError.

pub fn create_table_of_contents(&mut self) -> Result<u32, PdfError>

Creates the table of contents page(s).

Returns

Result<u32, PdfError>: Unit type or a propagated PdfError.

pub fn create_data_types_page( &mut self, column_types: &IndexMap<String, DataType>, ) -> Result<(), PdfError>

Creates the column type overview page.

Parameters

column_types: The index map of the column names and corresponding data types.

Returns

Result<(), PdfError>: Unit type or the propagated PdfError.

pub fn create_descriptive_analysis_page( &mut self, descriptive_analysis: &DescriptiveAnalysis, ) -> Result<(), LeadsError>

Creates the report pages with the results of the basic descriptie analysis.

Parameters

descriptive_analysis: The descriptive analysis results.

Returns

Result<() PdfError>: Unit type or the propogated PdfError.

pub fn create_missing_values_page( &mut self, missing_values_analysis: &MissingValueAnalysis, visualizations: &Option<VisualizationManager>, ) -> Result<(), PdfError>

Create the missing values analysis pages.

pub fn create_glossary_page(&mut self) -> Result<(), PdfError>

Creates the term glossary pages.

pub fn save_to_file(&self, path: &PathBuf) -> Result<(), PdfError>

Saves the document to disk.

Parameters

path: Path to save the file.

Returns

Result<(), PdfError>: Unit type of a propagated PdfError.